% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi_functions.R
\name{plot.HDIdata}
\alias{plot.HDIdata}
\title{Plot HDI Data}
\usage{
\method{plot}{HDIdata}(object)
}
\arguments{
\item{object}{An object of class 'HDIdata'.}
}
\value{
A ggplot object showing the trend of the indicators over time.
}
\description{
This function plots the trend of 4 indicators over time.
}
\examples{
plot.HDIdata(hdi_data)
}
