% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi_functions.R
\name{print.HDIdata}
\alias{print.HDIdata}
\title{Print HDIdata}
\usage{
\method{print}{HDIdata}(x)
}
\arguments{
\item{x}{An object of class 'HDIdata'.}
}
\value{
Prints a summary of the HDIdata object.
}
\description{
This method prints a summary of an HDIdata object, including the number of observations,
column names, country, indicators, and categories which the indicators belong to.
}
\examples{
print(hdi_data)
}
