% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi_functions.R
\name{read_HDI}
\alias{read_HDI}
\title{Read HDI Data}
\usage{
read_HDI(file_name)
}
\arguments{
\item{file_name}{The name of the file containing HDI data.}
}
\value{
An object of class 'HDIdata' with the data read from the file.
}
\description{
This function reads in HDI (Human Development Index) data from a specified file,
assigns appropriate column types, and returns an object of class 'HDIdata'.
}
\examples{
hdi_data <- read_HDI("path/to/your/hdi_data_file.csv")
}
